backend/services/ai-service/requirements.txt
fastapi>=0.110
uvicorn>=0.29
python-multipart>=0.0.9
torch>=2.0
torchvision>=0.15
timm>=1.0
Pillow>=10.0
numpy>=1.24
python-dotenv>=1.0
pydantic>=2.0
